Add timeout
Add timeout for login function, exclude existing entities in the same day

+++ irregular_tax.py
@@ -17,17 +17,19 @@ from PIL import Image
 import baidu_api
 import requests
 import time
+from func_timeout import func_timeout, FunctionTimedOut
 import utility_email as em
 from utility_commons import *
 
-SCREENSHOT_PATH = PIC_DIR + r'\screen_shot.jpg'
-VCODE_PATH = PIC_DIR + r'\vcode.jpg'
+SCREENSHOT_PATH = PIC_DIR + r'\screen_shot.png'
+VCODE_PATH = PIC_DIR + r'\vcode.png'
 FILE_PATH = FILE_DIR + r'\Irregular_Tax.xls'
 ATTACHMENT_PATH = FILE_DIR + r'\{}_异常发票清单_{}.xlsx'
 
 SITE = 'Irregular_Tax'
 TABLE_NAME = 'Scrapy_' + SITE
 ACCESS_TABLE_NAME = 'Scrapy_Irregular_Tax_Access'
+LOGS_TABLE_NAME = 'Scrapy_Logs'
 LOG_PATH = LOG_DIR + '\\' + SITE + '.log'
 
 logger = getLogger(SITE)
@@ -73,8 +75,12 @@ class Tax:
 
     # Not return validation code until get valid one
     def get_vcode(self):
-
+        count = 0
         while True:
+
+            logger.info('Try {} times.'.format(count))
+            count += 1
+
             vpic = self.get_vcode_pic()
             if vpic:
                 ocr = baidu_api.Baidu(api='ocr')
@@ -100,6 +106,7 @@ class Tax:
     # Login
     def login(self, username, password):
         while True:
+
             vcode = self.get_vcode()
             if vcode:
                 try:
@@ -165,7 +172,12 @@ class Tax:
     def run(cls, entity, server, link, username, password):
         t = cls(link, username, password)
         while True:
-            t.login(username=username, password=password)
+            try:
+                func_timeout(3000, t.login, args=(username, password))
+            except FunctionTimedOut:
+                logger('Login timeout.')
+                exit(1)
+            # t.login(username=username, password=password)
             success = t.get()
             if success:
                 t.web.close()
@@ -204,11 +216,19 @@ def _send_email(entity, receiver, attachment):
 
 if __name__ == '__main__':
 
+    logger.info('---------------   Irregular tax ratio query.   ---------------')
+
     with db.Mssql(keys.dbconfig) as scrapydb:
         access = scrapydb.select(ACCESS_TABLE_NAME)
-        logger.info('---------------   Irregular tax ratio query.   ---------------')
-        logger.info('Delete existing records.')
-        scrapydb.delete(TABLE_NAME)
+        entities = '\'' + '\', \''.join(list(access['Entity_Name'])) + '\''
+        logs = scrapydb.select(LOGS_TABLE_NAME, source=SITE, customized={'Timestamp': ">='{}'".format(TODAY), 'City': 'IN ({})'.format(entities)})
+        # Exclude entities with logs in same day. If no logs, refresh table
+        if not logs.empty:
+            logger.info('Exclude existing entities.')
+            access = access[-access['Entity_Name'].isin(logs['City'])]
+        else:
+            logger.info('Delete existing records.')
+            scrapydb.delete(TABLE_NAME)
 
     # Core scraping process
     for index, row in access.iterrows():
@@ -216,7 +236,7 @@ if __name__ == '__main__':
         logger.info('---------------   Start new job. Entity: {} Server:{}    ---------------'.format(row['Entity_Name'], row['Server']))
         result = Tax.run(entity=row['Entity_Name'], server=row['Server'], link=row['Link'], username=row['User_Name'], password=row['Password'])
         # Upload to database
-        scrapydb.upload(result, TABLE_NAME, False, False, None, start=PRE3MONTH, end=TODAY,  timestamp=TIMESTAMP)
+        scrapydb.upload(result, TABLE_NAME, False, False, None, start=PRE3MONTH, end=TODAY,  timestamp=TIMESTAMP, source=SITE, city=row['Entity_Name'])
         scrapydb.close()
         # time.sleep(20)
 
@@ -235,7 +255,7 @@ if __name__ == '__main__':
     # Send email summary
     # with em.Email() as scrapyemail_summary:
     scrapyemail_summary = em.Email()
-    scrapyemail_summary.send(SITE, 'Done', LOG_PATH, receivers='benson.chen@ap.jll.com;helen.hu@ap.jll.com')
+    scrapyemail_summary.send(TABLE_NAME, 'Done', LOG_PATH, receivers='benson.chen@ap.jll.com;helen.hu@ap.jll.com')
     scrapyemail_summary.close()
     exit()
